Delete Validations and Interface Changes

+++ frontend/src/components/QAndASection/DeleteAnswer.js
@@ -3,6 +3,7 @@ import { useHistory, useRouteMatch } from 'react-router-dom'
 import { deleteAnswer, getAnswers } from '../../api/apiFBQA';
 import { DeleteAnswerForm } from '../QAndASection/Forms/DeleteAnswerForm';
 import { useSelector } from "react-redux";
+import Swal from "sweetalert2";
 // import deleteAnswerBg from './ImagesD/deleteAnswerBg.png';
 
 export const DeleteAnswer = () =>{
@@ -23,8 +24,19 @@ export const DeleteAnswer = () =>{
     },[]);
 
     const onSubmit = async(data) =>{
-        await deleteAnswer(data, match.params.id)
-        history.push(`/member/myAnswers/${member_email}`);
+        Swal.fire({
+            title: 'Confirm the delete process!',
+            showCancelButton: true,
+            confirmButtonText: 'Delete',
+          }).then((result) => {
+            if (result.isConfirmed) {
+                deleteAnswer(data, match.params.id)
+                    Swal.fire('The Answer has been deleted!', '', 'success')
+                    history.push(`/member/myAnswers/${member_email}`);
+            } else {
+              Swal.fire('The delete process has been canceled!', '', 'info')
+            }
+          })  
     }
 
     return answer ?(
